FEAT: Per-pixel job allocation.
The previous algorithm divvied up the work inefficiently.  The previous
implementation divided the base rendermap into segments and then allocated
a thread to solve all the points in that segment.  Some segments, however,
may have no valid points at all; the orbits described may terminate early,
while other segments may be full of points with long orbits.

Each pixel on the base rendermap represents the starting point of an
orbit.  By testing and rendering each starting point individually, we
create a highly efficient queue; no CPU is ever idled during the rendering
of the final image.  There may be some loss of efficiency locking and
unlocking the job queue, especially in the corners of the map where there
will be very few valid orbits, and it may make sense to try and create a
work-stealing algorithm instead, but trying to map and split generators
(and Ranges are generators) to the crossbeam::Injector model feels like
it would be a bit of a nightmare.

this commit will...) <subject> (Max 50 char) # |<----  Using a Maximum
Of 50 Characters  ---->|