FEAT I give up
At this point, I'm beating a dead horse trying to figure out what's wrong with my
maths, and while I could keep trying, I've realized that I'm overworking this
problem to death and have better things to do.  This project is back-burnered until
I have an infinite amount of time to deal with it.